refacotr code
gitlab api client split to module
fixes #16
