+++ lib/git/gitlab/api/authorize.rb
@@ -1,0 +1,13 @@
+require "gitlab"
+require "grit"
+require "git/error"
+
+class GitlabApi::ApiClient
+	
+	module Authorize
+		def authorize
+			@client.user
+		end
+	end
+	
+end
